LayoutTests/inspector/canvas/shaderProgram-add-remove-webgl2.html

<!DOCTYPE html>
<html>
<head>
<script src="../../http/tests/inspector/resources/inspector-test.js"></script>
<script src="resources/shaderProgram-utilities-webgl.js"></script>
<script>
function test() {
    let suite = initializeTestSuite("webgl2");

    addSimpleTestCase("webgl2");
    addParentCanvasRemovedTestCase("webgl2");

    suite.runTestCasesAndFinish();
}
</script>
</head>
<body onload="runTest()">
    <p>Test that CanvasManager tracks creation and destruction of WebGL2 shader programs.</p>
</body>
<script>
    createProgram("webgl2");
</script>
</html>
